Guards added

+++ src/events/events.controller.ts
@@ -11,6 +11,8 @@ import {
   Query,
   UsePipes,
   ValidationPipe,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { EventsService } from './events.service';
 import { CreateEventDto } from './dto/create-event.dto';
@@ -20,6 +22,9 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Event } from './entities/event.entity';
 import { Attendee } from 'src/attendees/entities/attendee.entity';
 import { ListEvents } from './dto/lists-events';
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { User } from 'src/auth/entities/user.entity';
+import { AuthGuard } from '@nestjs/passport';
 
 @Controller('events')
 export class EventsController {
@@ -33,11 +38,12 @@ export class EventsController {
   ) {}
 
   @Post()
-  async create(@Body() createEventDto: CreateEventDto) {
-    return await this.eventRepo.save({
-      ...createEventDto,
-      when: new Date(createEventDto.when),
-    });
+  @UseGuards(AuthGuard('jwt'))
+  async create(
+    @Body() createEventDto: CreateEventDto,
+    @CurrentUser() user: User,
+  ) {
+    return await this.eventsService.createEvent(createEventDto, user);
   }
 
   @Get()
@@ -113,32 +119,43 @@ export class EventsController {
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuard('jwt'))
   async update(
     @Param('id') id: string,
     @Body() updateEventDto: UpdateEventDto,
+    @CurrentUser() user: User,
   ) {
-    const event = await this.eventRepo.findOneBy({ id: +id });
+    const event = await this.eventsService.getEvent(+id);
 
+    // check if event exists
     if (!event) {
       throw new NotFoundException();
     }
 
-    return await this.eventRepo.save({
-      ...event,
-      ...updateEventDto,
-      when: updateEventDto.when ? new Date(updateEventDto.when) : event.when,
-    });
+    // check user is the owner of the event
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(null, `You're not the owner of this event`);
+    }
+
+    return await this.eventsService.updateEvent(event, updateEventDto);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuard('jwt'))
   @HttpCode(204)
-  async remove(@Param('id') id: string) {
-    const result = await this.eventsService.deleteEvent(+id);
+  async remove(@Param('id') id: string, @CurrentUser() user: User) {
+    const event = await this.eventRepo.findOneBy({ id: +id });
 
-    if (result.affected !== 1) {
+    // check if event exists
+    if (!event) {
       throw new NotFoundException();
     }
 
-    return result;
+    // check user is the owner of the event
+    if (event.organizerId !== user.id) {
+      throw new ForbiddenException(null, `You're not the owner of this event`);
+    }
+
+    await this.eventsService.deleteEvent(+id);
   }
 }
